+++ cmd/wasm/logic/logic_test.go
@@ -1,16 +1,12 @@
-//go:build js
-// +build js
-
-package main
+package logic
 
 import (
-	"syscall/js"
 	"testing"
 )
 
 func Test_run(t *testing.T) {
 	type args struct {
-		args []js.Value
+		r, g, b float64
 	}
 	tests := []struct {
 		name     string
@@ -19,36 +15,27 @@ func Test_run(t *testing.T) {
 		wantErr  bool
 	}{
 		{
-			name: "happy path: Black",
-			args: args{
-				[]js.Value{js.ValueOf(0), js.ValueOf(0), js.ValueOf(0)},
-			},
+			name:     "happy path: Black",
+			args:     args{0, 0, 0},
 			wantHtml: `<div><label for="output_name" id="output_label">Color Name:</label><output name="color_name" id="output_name"> Black</output></div><div><label for="output_type" id="output_label">Color Type:</label><output name="color_type" id="output_type"> Cool</output></div>`,
 			wantErr:  false,
 		},
 		{
-			name: "happy path: Red",
-			args: args{
-				[]js.Value{js.ValueOf(255), js.ValueOf(0), js.ValueOf(0)},
-			},
+			name:     "happy path: Red",
+			args:     args{255, 0, 0},
 			wantHtml: `<div><label for="output_name" id="output_label">Color Name:</label><output name="color_name" id="output_name"> Red</output></div><div><label for="output_type" id="output_label">Color Type:</label><output name="color_type" id="output_type"> Warm</output></div>`,
 			wantErr:  false,
 		},
 		{
-			name:    "unhappy path: not enough input args",
-			args:    args{[]js.Value{js.ValueOf(255)}},
-			wantErr: true,
-		},
-		{
 			name:    "unhappy path: wrong input enough input args",
-			args:    args{[]js.Value{js.ValueOf(-1), js.ValueOf(0), js.ValueOf(0)}},
+			args:    args{-1, 0, 0},
 			wantErr: true,
 		},
 	}
 	for _, tt := range tests {
 		t.Run(
 			tt.name, func(t *testing.T) {
-				gotHtml, err := run(tt.args.args)
+				gotHtml, err := Run(tt.args.r, tt.args.g, tt.args.b)
 				if (err != nil) != tt.wantErr {
 					t.Errorf("run() error = %v, wantErr %v", err, tt.wantErr)
 					return
